Crate librespot_discovery

Advertises this device to Spotify clients in the local network.

This device will show up in the list of “available devices”. Once it is selected from the list, Credentials are received. Those can be used to establish a new Session with librespot_core.

This library uses mDNS and DNS-SD so that other devices can find it, and spawns an http server to answer requests of Spotify clients.

Structs

Builder

A builder for Discovery.

Credentials

Credentials to be used in librespot. The credentials are used to log into the Spotify API.

Discovery

Makes this device visible to Spotify clients in the local network.

Error

Enums

DeviceType

Determining the icon in the list of available devices.

DiscoveryError

Errors that can occur while setting up a Discovery instance.